Moving things around
Getting things out of _gunicorn.
AppConfig as the generic config for apps/workloads
