+++ paas_app_charmer/django/charm.py
@@ -101,7 +101,7 @@ class Charm(GunicornBase):  # pylint: disable=too-many-instance-attributes
                     **self._gen_environment(),
                 },
                 combine_stderr=True,
-                working_dir=str(self._workload_config.app_dir),
+                working_dir=str(self._app_config.app_dir),
             ).wait_output()
             event.set_results({"password": password})
         except ops.pebble.ExecError as e:
